+++ json-fleece-aeson-beeline/json-fleece-aeson-beeline.cabal
@@ -37,7 +37,7 @@ library
     , beeline-http-client >=0.2 && <0.9
     , bytestring ==0.11.*
     , http-client ==0.7.*
-    , json-fleece-aeson >=0.1 && <0.4
+    , json-fleece-aeson >=0.1 && <0.5
   default-language: Haskell2010
   if flag(strict)
     ghc-options: -Weverything -Werror -Wno-missing-local-signatures -Wno-monomorphism-restriction -Wno-missing-kind-signatures -Wno-prepositive-qualified-module -Wno-implicit-prelude -Wno-safe -Wno-unsafe -Wno-missing-safe-haskell-mode -Wno-missing-deriving-strategies -Wno-all-missed-specialisations -Wno-missed-specialisations -Wno-unticked-promoted-constructors
