+++ json-fleece-aeson-beeline/package.yaml
@@ -49,10 +49,10 @@ when:
 library:
   source-dirs: src
   dependencies:
-    - json-fleece-aeson >= 0.1 && < 0.4
     - beeline-http-client >= 0.2 && < 0.9
     - bytestring >= 0.11 && < 0.12
     - http-client >= 0.7 && < 0.8
+    - json-fleece-aeson >= 0.1 && < 0.5
   exposed-modules:
     - Fleece.Aeson.Beeline
