+++ json-fleece-aeson/package.yaml
@@ -1,5 +1,5 @@
 name:                json-fleece-aeson
-version:             0.3.5.0
+version:             0.4.0.0
 github:              "flipstone/json-fleece/json-fleece-aeson"
 license:             BSD3
 author:              "Author name here"
@@ -16,14 +16,14 @@ copyright:           "2023 Author name here"
 description:         Please see the README on GitHub at <https://github.com/githubuser/json-fleece-aeson#readme>
 
 dependencies:
-- base >= 4.7 && < 5
 - aeson >= 2.0 && < 2.2
+- base >= 4.7 && < 5
 - bytestring >= 0.11 && < 0.12
 - containers >= 0.6 && < 0.7
-- json-fleece-core >= 0.7 && < 0.8
+- json-fleece-core >= 0.8 && < 0.9
+- shrubbery >= 0.2 && < 0.3
 - text >= 1.2 && < 2.1
 - vector >= 0.12 && < 0.14
-- shrubbery >= 0.2 && < 0.3
 
 flags:
   strict:
@@ -58,6 +58,8 @@ library:
     - Fleece.Aeson
     - Fleece.Aeson.Decoder
     - Fleece.Aeson.Encoder
+  dependencies:
+    - json-fleece-openapi3 >= 0.5 && < 0.6
 
 tests:
   json-fleece-aeson-test:
