+++ json-fleece-core/json-fleece-core.cabal
@@ -5,7 +5,7 @@ cabal-version: 1.12
 -- see: https://github.com/sol/hpack
 
 name:           json-fleece-core
-version:        0.7.0.0
+version:        0.8.0.0
 description:    Please see the README on GitHub at <https://github.com/flipstone/json-fleece-core#readme>
 homepage:       https://github.com/flipstone/json-fleece#readme
 bug-reports:    https://github.com/flipstone/json-fleece/issues
@@ -33,6 +33,7 @@ library
       Fleece.Core.Name
       Fleece.Core.Schemas
   other-modules:
+      Fleece.Core.Validator
       Paths_json_fleece_core
   hs-source-dirs:
       src
