+++ json-fleece-examples/package.yaml
@@ -1,5 +1,5 @@
 name:                json-fleece-examples
-version:             0.2.3.0
+version:             0.3.0.0
 github:              "flipstone/json-fleece/json-fleece-examples"
 license:             BSD3
 author:              "Author name here"
@@ -18,7 +18,7 @@ description:         Please see the README on GitHub at <https://github.com/flip
 dependencies:
 - base >= 4.7 && < 5
 - containers >= 0.6 && < 0.7
-- json-fleece-core >= 0.7 && < 0.8
+- json-fleece-core >= 0.8 && < 0.9
 - scientific >= 0.3.7 && < 0.4
 - shrubbery >= 0.2 && < 0.3
 - text >= 1.2 && < 2.1
