+++ json-fleece-hermes/package.yaml
@@ -1,5 +1,5 @@
 name:                json-fleece-hermes
-version:             0.1.1.0
+version:             0.2.0.0
 github:              "flipstone/json-fleece/json-fleece-hermes"
 license:             BSD3
 author:              "Author name here"
@@ -18,7 +18,7 @@ description: Please see the README on GitHub at <https://github.com/flipstone/js
 dependencies:
 - base >= 4.7 && < 5
 - bytestring >= 0.11 && < 0.12
-- json-fleece-core >= 0.7 && < 0.8
+- json-fleece-core >= 0.8 && < 0.9
 - text >= 2.0 # This could be problematic for anyone using LTS 20 or below
 
 flags:
@@ -55,6 +55,7 @@ library:
   dependencies:
     - containers >= 0.6 && < 0.7
     - hermes-json >= 0.6 && < 0.7
+    - json-fleece-openapi3 >= 0.5 && < 0.6
     - shrubbery >= 0.2 && < 0.3
 
 tests:
