+++ json-fleece-markdown/package.yaml
@@ -1,5 +1,5 @@
 name:                json-fleece-markdown
-version:             0.5.0.0
+version:             0.6.0.0
 github:              "flipstone/json-fleece/json-fleece-markdown"
 license:             BSD3
 author:              "Author name here"
@@ -52,9 +52,10 @@ library:
   exposed-modules:
     - Fleece.Markdown
   dependencies:
-    - json-fleece-core >= 0.7 &&  < 0.8
     - containers >= 0.6 && < 0.7
     - dlist >= 1.0 && < 1.1
+    - json-fleece-core >= 0.8 && < 0.9
+    - json-fleece-openapi3 >= 0.5 && < 0.6
 
 tests:
   json-fleece-markdown-test:
