+++ json-fleece-openapi3/examples/star-trek/star-trek.cabal
@@ -1870,7 +1870,7 @@ library
     , beeline-http-client ==0.8.*
     , beeline-routing >=0.2.4 && <0.3
     , json-fleece-aeson-beeline ==0.2.*
-    , json-fleece-core >=0.1.3 && <0.8
+    , json-fleece-core ==0.8.*
     , scientific
     , text
     , time
