+++ json-fleece-openapi3/examples/test-cases/test-cases.cabal
@@ -198,7 +198,7 @@ library
     , beeline-routing >=0.2.4 && <0.3
     , containers
     , json-fleece-aeson-beeline ==0.2.*
-    , json-fleece-core >=0.1.3 && <0.8
+    , json-fleece-core ==0.8.*
     , scientific
     , shrubbery ==0.2.*
     , text
