+++ json-fleece-openapi3/package.yaml
@@ -1,5 +1,5 @@
 name:                json-fleece-openapi3
-version:             0.4.3.1
+version:             0.5.0.0
 github:              "flipstone/json-fleece/json-fleece-openapi3"
 license:             BSD3
 author:              "Author name here"
@@ -61,12 +61,16 @@ library:
     - Fleece.OpenApi3
   dependencies:
     - aeson >= 2.0 && < 2.2
-    - insert-ordered-containers >= 0.2 && < 0.3
     - containers >= 0.6 && < 0.7
+    - insert-ordered-containers >= 0.2 && < 0.3
+    - json-fleece-core >= 0.8 && < 0.9
     - mtl >= 2.2 && < 2.4
     - non-empty-text >= 0.2 && < 0.3
     - openapi3 >= 3.2 && < 3.3
+    - scientific >= 0.3.7 && < 0.4
     - text >= 1.2 && < 2.1
+    - time >= 1.11 && < 1.13
+    - vector >= 0.12 && < 0.14
 
 executables:
   fleece-openapi3:
