+++ json-fleece-openapi3/src/Fleece/OpenApi3/Schemas.hs
@@ -1,0 +1,6 @@
+{-# OPTIONS_GHC -Wno-missing-import-lists #-}
+
+module Fleece.OpenApi3.Schemas (module Export) where
+
+import Fleece.OpenApi3.Schemas.OpenApi3Validator as Export
+import Fleece.OpenApi3.Schemas.Schemas as Export
