+++ json-fleece-pretty-print/json-fleece-pretty-print.cabal
@@ -5,7 +5,7 @@ cabal-version: 1.12
 -- see: https://github.com/sol/hpack
 
 name:           json-fleece-pretty-print
-version:        0.1.3.0
+version:        0.2.0.0
 description:    Please see the README on GitHub at <https://github.com/flipstone/json-fleece-pretty-print#readme>
 homepage:       https://github.com/flipstone/json-fleece#readme
 bug-reports:    https://github.com/flipstone/json-fleece/issues
@@ -36,7 +36,8 @@ library
       base >=4.7 && <5
     , containers ==0.6.*
     , dlist ==1.0.*
-    , json-fleece-core ==0.7.*
+    , json-fleece-core ==0.8.*
+    , json-fleece-openapi3 ==0.5.*
     , scientific ==0.3.*
     , shrubbery ==0.2.*
     , text >=1.2 && <2.1
@@ -58,7 +59,7 @@ test-suite json-fleece-pretty-print-test
       base >=4.7 && <5
     , containers ==0.6.*
     , hedgehog
-    , json-fleece-core ==0.7.*
+    , json-fleece-core ==0.8.*
     , json-fleece-examples
     , json-fleece-pretty-print
     , shrubbery ==0.2.*
