+++ json-fleece-pretty-print/package.yaml
@@ -1,5 +1,5 @@
 name:                json-fleece-pretty-print
-version:             0.1.3.0
+version:             0.2.0.0
 github:              "flipstone/json-fleece/json-fleece-pretty-print"
 license:             BSD3
 author:              "Author name here"
@@ -17,7 +17,7 @@ description:         Please see the README on GitHub at <https://github.com/flip
 
 dependencies:
 - base >= 4.7 && < 5
-- json-fleece-core >= 0.7 && < 0.8
+- json-fleece-core >= 0.8 && < 0.9
 - shrubbery >= 0.2 && < 0.3
 - text >= 1.2 && < 2.1
 
@@ -55,6 +55,7 @@ library:
   dependencies:
     - containers >= 0.6 && < 0.7
     - dlist >= 1.0 && < 1.1
+    - json-fleece-openapi3 >= 0.5 && < 0.6
     - scientific >= 0.3 &&  < 0.4
 
 tests:
