+++ json-fleece-swagger2/examples/uber/package.yaml
@@ -13,7 +13,7 @@ dependencies:
   - base >= 4.7 && < 5
   - text
   - scientific
-  - json-fleece-core >= 0.1.3 && < 0.8
+  - json-fleece-core >= 0.8 && < 0.9
   - json-fleece-aeson-beeline >= 0.2 && < 0.3
   - beeline-routing >= 0.2.4 && < 0.3
   - beeline-http-client >= 0.8 && < 0.9
