+++ json-fleece-swagger2/examples/uber/uber.cabal
@@ -73,7 +73,7 @@ library
     , beeline-http-client ==0.8.*
     , beeline-routing >=0.2.4 && <0.3
     , json-fleece-aeson-beeline ==0.2.*
-    , json-fleece-core >=0.1.3 && <0.8
+    , json-fleece-core ==0.8.*
     , scientific
     , text
     , time
