+++ json-fleece-swagger2/json-fleece-swagger2.cabal
@@ -5,7 +5,7 @@ cabal-version: 1.12
 -- see: https://github.com/sol/hpack
 
 name:           json-fleece-swagger2
-version:        0.4.0.4
+version:        0.5.0.0
 description:    Please see the README on GitHub at <https://github.com/githubuser/json-fleece-swagger2#readme>
 homepage:       https://github.com/flipstone/json-fleece#readme
 bug-reports:    https://github.com/flipstone/json-fleece/issues
@@ -86,7 +86,7 @@ library
       base >=4.7 && <5
     , insert-ordered-containers ==0.2.*
     , json-fleece-codegen-util >=0.6 && <0.11
-    , json-fleece-openapi3 ==0.4.*
+    , json-fleece-openapi3 ==0.5.*
     , openapi3 ==3.2.*
     , swagger2 ==2.8.*
     , text >=1.2 && <2.1
